feat: hook to retrieve exchange listings for coin
Adds a hook to retrieve exchange listings for a given coin. Includes unit tests to verify functionality. This hook along with useExchangePairs hook will almost certainly be used outside of the transaction form. Refactoring to move to a common location will be done when they're needed outside of the transaction form.